+++ Examples/ISHDeploy/Set-ISHSTSWindows.ImplicitRemoting.ps1
@@ -26,7 +26,7 @@ param (
 )      
 $ishBootStrapRootPath=Resolve-Path "$PSScriptRoot\..\.."
 $cmdletsPaths="$ishBootStrapRootPath\Source\Cmdlets"
-$scriptsPaths="$ishBootStrapRootPath\Source\Scripts"
+$serverScriptsPaths="$ishBootStrapRootPath\Source\Server"
 
 . $ishBootStrapRootPath\Examples\ISHDeploy\Cmdlets\Write-Separator.ps1
 Write-Separator -Invocation $MyInvocation -Header -Name "Configure"
@@ -35,7 +35,7 @@ $scriptProgress=Get-ProgressHash -Invocation $MyInvocation
 
 if(-not $Computer)
 {
-    & "$scriptsPaths\Helpers\Test-Administrator.ps1"
+    & "$serverScriptsPaths\Helpers\Test-Administrator.ps1"
 }
 
 . $cmdletsPaths\Helpers\Add-ModuleFromRemote.ps1